UPDATED: arm64: introduce CPU_OF_TABLES for cpu ops selection
Add support to arm64 to provide a dt-based method to
allow soc-vendors to supply cpu_ops. Also move psci and
smp_spin_table ops to use CPU_OF_TABLES.

Change-Id: I4fd5b330fa3a7436fb88fff385deeb80a90548d6
[abhimany: ignore the upstream changes for specifying the
cpu ops in code and continue to let it be defined in DT in
order to allow co-existence of PSCI and msm defined cpu ops]
Signed-off-by: Abhimanyu Kapur <abhimany@codeaurora.org>

This patch is modified to take account of the ARM64 ACPI
Parking Protocol specification, implemented in current kernel
(k4.4).

Signed-off-by: AngeloGioacchino Del Regno <kholk@gentoo.org>
Signed-off-by: Pavel Dubrova <pashadubrova@gmail.com>